api/pyproject.toml
[tool.poetry]
name = "gymlog-server"
version = "1.1.4"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10,<3.12"
fastapi = "^0.95.1"
uvicorn = { extras = ["standard"], version = "^0.22.0" }
python-dotenv = "^1.0.0"
sqlalchemy = "^1.4.48"
mysqlclient = "^2.1.1"
pydantic = "^1.10.7"
sse-starlette = "^1.6.1"
alembic = "^1.11.1"
gunicorn = "^20.1.0"
